First step towards updates without a tarball directory
Will NOT work with unreleased tarballs.

+++ manage_repos/manage_repos.py
@@ -204,8 +204,8 @@ def update_version(specfile: str, version_to: str, patches=None) -> None:
             print(line)
 
 
-def update_package(entry: Path, version_to: str, tarball_directory: Path, *,
-                   committer: str, kind: str="applications",
+def update_package(entry: Path, version_to: str, tarball_directory: Path=None,
+                   *, committer: str, kind: str="applications",
                    changetype: str="bugfix", checkout_dir: Path=None,
                    upstream_branch: str=None) -> bool:
 
@@ -227,24 +227,36 @@ def update_package(entry: Path, version_to: str, tarball_directory: Path, *,
     tarball_name = "{name}-{version_to}.tar.xz".format(name=upstream_reponame,
                                                        version_to=version_to)
 
-    done_subdir = tarball_directory / "done"
-    done_subdir.mkdir(exist_ok=True)
-
-    if (done_subdir / tarball_name).exists():
-        print("Tarball {} already processed, skipping".format(tarball_name))
-        return False
-
-    tarball_path = tarball_directory / tarball_name
-    destination_path = entry / tarball_name
-
-    if not tarball_path.exists():
-        print("Tarball {} missing, skipping".format(tarball_name))
-        return False
-
-    shutil.copy(str(tarball_path), str(destination_path))
-    tarball_path.rename(done_subdir / tarball_name)
-
     update_version(specfile, version_to, patches)
+
+    if tarball_directory is not None and tarball_directory.exists():
+        done_subdir = tarball_directory / "done"
+        done_subdir.mkdir(exist_ok=True)
+
+        if (done_subdir / tarball_name).exists():
+            print("Tarball {} already processed, skipping".format(
+                tarball_name))
+            return False
+
+        tarball_path = tarball_directory / tarball_name
+        destination_path = entry / tarball_name
+
+        if not tarball_path.exists():
+            print("Tarball {} missing, skipping".format(tarball_name))
+            return False
+
+        shutil.copy(str(tarball_path), str(destination_path))
+        tarball_path.rename(done_subdir / tarball_name)
+
+    elif not (entry / tarball_name).exists():
+        # Try to download the tarball if not present
+        print("No tarball found. Attempting download...")
+        cmd = ["osc", "service", "localrun", "download_files"]
+        result = run(cmd)
+        if result.returncode != 0:
+            print("Download of {} failed. Skipping package.".format(
+                tarball_name))
+            return False
 
     record_changes(changes_file, checkout_dir, current_version,
                    version_to, upstream_reponame=upstream_reponame,
